refactor(session-persistence): extract deletion owner
